[GREEN] REFACTOR: Fix duplication issue on EditPond and AddPond
