+++ components/pond/AddPond.tsx
@@ -1,13 +1,12 @@
 'use client';
 
-import React, { useState, useEffect } from 'react';
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
+import React, { useState } from 'react';
 import { Button } from '../ui/button';
-import { PondAddSchema, PondAddForm } from '@/types/pond/addpond';
 import { addPond } from '@/lib/pond';
 import { IoIosAdd } from 'react-icons/io';
 import { Modal } from '../ui/modal';
+import PondForm from './PondForm';
+import { PondInputForm } from '@/types/pond/inputpond';
 
 interface AddPondProps extends React.HTMLProps<HTMLDivElement> {
   token?: string;
@@ -16,42 +15,21 @@ interface AddPondProps extends React.HTMLProps<HTMLDivElement> {
 const AddPond: React.FC<AddPondProps> = ({ token: propToken, ...props }) => {
   const [open, setOpen] = useState(false);
   const [error, setError] = useState<string | null>(null);
-  const [volume, setVolume] = useState<number | null>(null);
 
-  const { register, handleSubmit, setValue, watch, formState: { errors }, reset } = useForm<PondAddForm>({
-    resolver: zodResolver(PondAddSchema),
-  });
-
-  const length = watch('length');
-  const width = watch('width');
-  const height = watch('depth');
-
-  useEffect(() => {
-    if (length && width && height) {
-      const calculatedVolume = parseFloat(length.toString()) * parseFloat(width.toString()) * parseFloat(height.toString());
-      setVolume(calculatedVolume);
-      setValue('volume', calculatedVolume); 
-    }
-  }, [length, width, height, setValue]);
-
-  const onSubmit = async (data: PondAddForm) => {
+  const onSubmit = async (data: PondInputForm) => {
     if (!propToken) {
       console.error('No token found');
       setError('No token found');
       return;
     }
-    console.log('Submitting data', data);
 
     try {
-      const response = await addPond(data, propToken);
-      console.log('Pond created:', response);
-      reset();
+      await addPond(data, propToken);
       setOpen(false);
-      console.log('Modal closed');
     } catch (error) {
       console.error('Failed to create pond:', error);
       setError('Failed to create pond');
-    } 
+    }
   };
 
   return (
@@ -62,68 +40,8 @@ const AddPond: React.FC<AddPondProps> = ({ token: propToken, ...props }) => {
 
       <Modal open={open} onClose={() => setOpen(false)}>
         <div role="dialog">
-          <form onSubmit={handleSubmit(onSubmit)} className='space-y-4'>
-            <input
-              {...register('name')}
-              placeholder='Nama Kolam'
-              className='w-full p-3 border border-gray-300 rounded-lg'
-            />
-            {errors.name && <p className='text-red-500'>{errors.name.message?.toString()}</p>}
-
-            <input
-              {...register('image_name')}
-              placeholder='Nama Gambar'
-              className='w-full p-3 border border-gray-300 rounded-lg'
-            />
-            {errors.image_name && <p className='text-red-500'>{errors.image_name.message?.toString()}</p>}
-
-            <input
-              {...register('length', {
-                setValueAs: value => parseFloat(value),
-              })}
-              placeholder='Panjang (meter)'
-              type='number'
-              className='w-full p-3 border border-gray-300 rounded-lg'
-            />
-            {errors.length && <p className='text-red-500'>{errors.length.message?.toString()}</p>}
-
-            <input
-              {...register('width', {
-                setValueAs: value => parseFloat(value),
-              })}
-              placeholder='Lebar (meter)'
-              type='number'
-              className='w-full p-3 border border-gray-300 rounded-lg'
-            />
-            {errors.width && <p className='text-red-500'>{errors.width.message?.toString()}</p>}
-
-            <input
-              {...register('depth', {
-                setValueAs: value => parseFloat(value),
-              })}
-              placeholder='Kedalaman (meter)'
-              type='number'
-              className='w-full p-3 border border-gray-300 rounded-lg'
-            />
-            {errors.depth && <p className='text-red-500'>{errors.depth.message?.toString()}</p>}
-
-            {volume !== null && (
-              <p className="text-gray-700">
-                Volume Kolam: {volume.toFixed(2)} m^3
-              </p>
-            )}
-
-            <input
-              type="hidden"
-              {...register('volume', { valueAsNumber: true })}
-            />
-
-            {error && <p className='text-red-500'>{error}</p>}
-
-            <Button type='submit'>
-              Submit
-            </Button>
-          </form>
+          <PondForm onSubmit={onSubmit} buttonText="Submit" />
+          {error && <p className='text-red-500'>{error}</p>}
         </div>
       </Modal>
     </div>
